requirements.txt
Flask>=3.0
Flask-JWT-Extended>=4.6
Flask-SQLAlchemy>=3.1
Werkzeug>=3.0
python-dotenv>=1.0
psycopg2-binary>=2.9
gunicorn>=21.0
